fix: rename Cloudinary helpers
