+++ app/(pages)/account-settings/mainContent.tsx
@@ -9,7 +9,7 @@ import { MdOutlineEmail } from 'react-icons/md';
 import { SlCalender } from 'react-icons/sl';
 import SelectAvatorModal from '@/components/modal/SelectAvatorModal';
 import { Spinner, useDisclosure } from '@nextui-org/react';
-// import { sendCloudinaryImage } from "@/util/SendCloudineryImage";
+// import { sendCloudinaryImage } from "@/util/SendCloudinaryImage";
 import { PhoneInput } from 'react-international-phone';
 import 'react-international-phone/style.css';
 import ProfileLoading from '@/components/loading/ProfileLoading';
@@ -20,7 +20,7 @@ import { format, parse } from 'date-fns';
 // import isUrl from "@/util/isUrl";
 import { updateUserProfile } from '@/actions/updateUserProfile';
 import { useRouter } from 'next/navigation';
-import { sendCloudinaryImage } from '@/lib/SendCloudineryImage';
+import { sendCloudinaryImage } from '@/lib/SendCloudinaryImage';
 import toast from 'react-hot-toast';
 import isUrl from '@/lib/isUrl';
 import UploadImageButton from '@/components/Button/UploadImageButton';
