+++ components/feed/PostFeed.tsx
@@ -12,8 +12,8 @@ import { postFeed } from "@/actions/postFeed";
 import { useUser } from "@/lib/UserContext";
 // import { useToast } from "@/hooks/use-toast";
 import DynamicPrimaryBtn from "../ui/Button/DynamicPrimaryBtn";
-import { sendCloudinaryImage } from "@/lib/SendCloudineryImage";
-import { sendCloudinaryVideo } from "@/lib/sendCloudineryVideo";
+import { sendCloudinaryImage } from "@/lib/SendCloudinaryImage";
+import { sendCloudinaryVideo } from "@/lib/sendCloudinaryVideo";
 import UserImageAvatar from "../util/Avatar";
 import isUrl from "@/lib/isUrl";
 import toast from "react-hot-toast";
