+++ components/smartsite/EditMicrosite/SwopPay/UpdateSwopPay.tsx
@@ -17,14 +17,14 @@ import { FaAngleDown, FaTimes } from "react-icons/fa";
 import { MdDelete, MdInfoOutline } from "react-icons/md";
 // import AnimateButton from "@/components/Button/AnimateButton";
 // import { deleteEmbedLink, updateEmbedLink } from "@/actions/embedLink";
-// import { sendCloudinaryVideo } from "@/util/sendCloudineryVideo";
+// import { sendCloudinaryVideo } from "@/util/sendCloudinaryVideo";
 // import { deleteVideo, postVideo, updateVideo } from "@/actions/video";
 // import placeholder from "@/public/images/video_player_placeholder.gif";
 // import CustomFileInput from "@/components/CustomFileInput";
-// import { sendCloudinaryImage } from "@/util/SendCloudineryImage";
+// import { sendCloudinaryImage } from "@/util/SendCloudinaryImage";
 import { deleteSwopPay, updateSwopPay } from "@/actions/swopPay";
 import { currencyItems, icon } from "@/components/util/data/smartsiteIconData";
-import { sendCloudinaryImage } from "@/lib/SendCloudineryImage";
+import { sendCloudinaryImage } from "@/lib/SendCloudinaryImage";
 import CustomFileInput from "@/components/CustomFileInput";
 import AnimateButton from "@/components/ui/Button/AnimateButton";
 import toast from "react-hot-toast";
