+++ components/smartsite/EditMicrosite/Video/AddVideo.tsx
@@ -9,9 +9,9 @@ import placeholder from "@/public/images/video_player_placeholder.gif";
 // import "react-quill/dist/quill.snow.css";
 // import CustomFileInput from "@/components/CustomFileInput";
 import { postVideo } from "@/actions/video";
-// import { sendCloudinaryVideo } from "@/util/sendCloudineryVideo";
+// import { sendCloudinaryVideo } from "@/util/sendCloudinaryVideo";
 import { FaTimes } from "react-icons/fa";
-import { sendCloudinaryVideo } from "@/lib/sendCloudineryVideo";
+import { sendCloudinaryVideo } from "@/lib/sendCloudinaryVideo";
 import CustomFileInput from "@/components/CustomFileInput";
 import AnimateButton from "@/components/ui/Button/AnimateButton";
 import toast from "react-hot-toast";
