+++ components/smartsite/EditMicrosite/Video/UpdateVideo.tsx
@@ -17,11 +17,11 @@ import { FaTimes } from "react-icons/fa";
 import { MdDelete, MdInfoOutline } from "react-icons/md";
 // import AnimateButton from "@/components/Button/AnimateButton";
 // import { deleteEmbedLink, updateEmbedLink } from "@/actions/embedLink";
-// import { sendCloudinaryVideo } from "@/util/sendCloudineryVideo";
+// import { sendCloudinaryVideo } from "@/util/sendCloudinaryVideo";
 import { deleteVideo, updateVideo } from "@/actions/video";
 // import placeholder from "@/public/images/video_player_placeholder.gif";
 // import CustomFileInput from "@/components/CustomFileInput";
-import { sendCloudinaryVideo } from "@/lib/sendCloudineryVideo";
+import { sendCloudinaryVideo } from "@/lib/sendCloudinaryVideo";
 import CustomFileInput from "@/components/CustomFileInput";
 import AnimateButton from "@/components/ui/Button/AnimateButton";
 import { Tooltip } from "@nextui-org/react";
