+++ components/smartsite/EditMicrosite/audio/AddAudio.tsx
@@ -9,12 +9,12 @@ import useSmartSiteApiDataStore from "@/zustandStore/UpdateSmartsiteInfo";
 // import CustomFileInput from "@/components/CustomFileInput";
 import AudioPlayer from "react-h5-audio-player";
 import "react-h5-audio-player/lib/styles.css";
-// import { sendCloudinaryImage } from "@/util/SendCloudineryImage";
+// import { sendCloudinaryImage } from "@/util/SendCloudinaryImage";
 import { postAudio } from "@/actions/audio";
-// import { sendCloudinaryAudio } from "@/util/sendCloudineryAudio";
+// import { sendCloudinaryAudio } from "@/util/sendCloudinaryAudio";
 import { FaTimes } from "react-icons/fa";
-import { sendCloudinaryAudio } from "@/lib/sendCloudineryAudio";
-import { sendCloudinaryImage } from "@/lib/SendCloudineryImage";
+import { sendCloudinaryAudio } from "@/lib/sendCloudinaryAudio";
+import { sendCloudinaryImage } from "@/lib/SendCloudinaryImage";
 import CustomFileInput from "@/components/CustomFileInput";
 import AnimateButton from "@/components/ui/Button/AnimateButton";
 import { MdInfoOutline } from "react-icons/md";
