+++ components/smartsite/EditMicrosite/audio/UpdateAudio.tsx
@@ -4,14 +4,14 @@ import { LiaFileMedicalSolid } from "react-icons/lia";
 import { FaTimes } from "react-icons/fa";
 // import AnimateButton from "@/components/Button/AnimateButton";
 import { MdDelete, MdInfoOutline } from "react-icons/md";
-// import { sendCloudinaryImage } from "@/util/SendCloudineryImage";
+// import { sendCloudinaryImage } from "@/util/SendCloudinaryImage";
 import Image from "next/image";
 // import CustomFileInput from "@/components/CustomFileInput";
 import { deleteAudio, updateAudio } from "@/actions/audio";
 import "react-h5-audio-player/lib/styles.css";
 import AudioPlayer from "react-h5-audio-player";
-import { sendCloudinaryAudio } from "@/lib/sendCloudineryAudio";
-import { sendCloudinaryImage } from "@/lib/SendCloudineryImage";
+import { sendCloudinaryAudio } from "@/lib/sendCloudinaryAudio";
+import { sendCloudinaryImage } from "@/lib/SendCloudinaryImage";
 import CustomFileInput from "@/components/CustomFileInput";
 import AnimateButton from "@/components/ui/Button/AnimateButton";
 import { Tooltip } from "@nextui-org/react";
