+++ components/smartsite/EditMicrosite/infoBar/AddInfoBar.tsx
@@ -30,7 +30,7 @@ import toast from "react-hot-toast";
 import Cookies from "js-cookie";
 import customImg from "@/public/images/IconShop/Upload@3x.png";
 import CustomFileInput from "@/components/CustomFileInput";
-import { sendCloudinaryImage } from "@/lib/SendCloudineryImage";
+import { sendCloudinaryImage } from "@/lib/SendCloudinaryImage";
 
 const AddInfoBar = ({ handleRemoveIcon, handleToggleIcon }: any) => {
   const [accessToken, setAccessToken] = useState("");
