+++ components/smartsite/EditMicrosite/infoBar/UpdateInfoBar.tsx
@@ -30,7 +30,7 @@ import Cookies from "js-cookie";
 import customImg from "@/public/images/IconShop/Upload@3x.png";
 
 import CustomFileInput from "@/components/CustomFileInput";
-import { sendCloudinaryImage } from "@/lib/SendCloudineryImage";
+import { sendCloudinaryImage } from "@/lib/SendCloudinaryImage";
 
 const UpdateInfoBar = ({ iconDataObj, isOn, setOff }: any) => {
   const state: any = useSmartSiteApiDataStore((state) => state); //get small icon store value
